Add Server
Server is the top level construct that ties various
components together.

It initializes informers and registers handlers for
the various objects. keeping track of changes of the various
objects as well as the current state of the cluster WRT
the tracked objects.

Main business logic is implemented in SyncMultiPolicy
method. which is invoked in a rate limited manner
to sync TC rules for representors currently associated
with pod interfaces scheduled on the current node

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>