Add builder to Ingress QDisc
Add builder to ingress qdisc to facilitate a way
to easily build ingress qdisc with parent and handle
QDiscAttrs.

In addition rename QDiscAttr to QdiscAttrs as a plural form
better reflects the type.

This change is needed for subsequent commits that add tests
to tc cmdline driver

unit tests are adjusted to the new API

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>