Change Flower filter DstIP to net.IPNet
currently Flower filter DstIP is a string that is expected
to be either an IP or CIDR wih special handling for transforming
an IP to CIDR. currently this logic was only applied to IPv4 addresses.
Which caused filters with IPv6 /128 mask to be re-applied on interfaces.

To solve this, we take a more structural approach for DstIP field
and use net.IPnet instead of plain string.

Signed-off-by: adrianc <adrianc@nvidia.com>